Fixes commented on pull request (new lines at end of files,  new comments and move method populate_standard_test)
