+++ app/automatization/scheduler_manager.py
@@ -7,6 +7,7 @@ from app.report.reports_generation import monthly_reports_generation
 MAX_NUMBER_OF_QUERIES = 1000
 
 
+# Job will be done the first day of every month
 @cron(0, 0, 1, -1, -1, target="mule")
 def reports_generation(num):
     """
@@ -18,6 +19,7 @@ def reports_generation(num):
     application.logger.info("Reports has been generated")
 
 
+# Job will be done at 3.00 am every day
 @cron(0, 3, -1, -1, -1, target="mule")
 def antennas_geolocalization(num):
     """
