+++ tests/events/save_wifi_event_test.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from app import application
-from app.data.populate_methods import populate_standard_test
+from tests.populate_db.populate_methods import populate_standard_test
 from app.models.wifi_traffic_event import WifiTrafficEvent
 from config import AppTokens
 from tests import base_test_case
