+++ tests/populate_db/initial_data_test.py
@@ -1,7 +1,7 @@
 from app import application
-from app.data.populate_methods import populate_standard_test
 from app.models.carrier import Carrier
 from tests import base_test_case
+from tests.populate_db.populate_methods import populate_standard_test
 
 
 class InitialdataTestCase(base_test_case.BaseTestCase):
